+++ qatrack/qa/views/perform.py
@@ -859,6 +859,7 @@ class ChooseUnit(TemplateView):
             q = q.values(
                 'unit',
                 'unit__type__name',
+                'unit__type__collapse',
                 'unit__name',
                 'unit__number',
                 'unit__id',
@@ -880,9 +881,11 @@ class ChooseUnit(TemplateView):
 
                 if unit['unit__site__name']:
                     key = (unit['unit__site__slug'], unit['unit__site__name'])
-                    unit_site_types[key][unit['unit__type__name']].append(unit)
+                    unit_site_types[key][(unit["unit__type__name"], unit["unit__type__collapse"])].append(unit)
                 else:
-                    unit_site_types[('zzzNonezzz', 'zzzNonezzz')][unit['unit__type__name']].append(unit)
+                    unit_site_types[('zzzNonezzz', 'zzzNonezzz')][
+                        (unit['unit__type__name'], unit['unit__type__collapse'])
+                    ].append(unit)
 
             ordered = {}
             for s in unit_site_types:
@@ -900,7 +903,7 @@ class ChooseUnit(TemplateView):
             context['split_by'] = int(split_by)
 
         else:
-            q = q.values('unit', 'unit__type__name', 'unit__name', 'unit__number',
+            q = q.values('unit', 'unit__type__name', 'unit__type__collapse', 'unit__name', 'unit__number',
                          'unit__id').order_by(units_ordering).distinct()
             freq_qs = models.Frequency.objects.prefetch_related('unittestcollections__unit').all()
 
@@ -910,7 +913,7 @@ class ChooseUnit(TemplateView):
                     'slug', 'name'
                 )
                 unit['categories'] = get_unit_categories(unit['unit__id'])
-                unit_types[unit["unit__type__name"]].append(unit)
+                unit_types[(unit["unit__type__name"], unit["unit__type__collapse"])].append(unit)
 
             ordered = sorted(list(unit_types.items()), key=lambda x: min([u[units_ordering] for u in x[1]]))
             context['split_sites'] = False